+++ app/src/main/java/ru/redbyte/exchangerate/domain/balance/SaveBalance.kt
@@ -3,6 +3,7 @@ package ru.redbyte.exchangerate.domain.balance
 import io.reactivex.Completable
 import ru.redbyte.exchangerate.data.exchange.Currency
 import ru.redbyte.exchangerate.domain.CompletableUseCase
+import java.math.BigDecimal
 import javax.inject.Inject
 
 class SaveBalance @Inject constructor(
@@ -11,4 +12,4 @@ class SaveBalance @Inject constructor(
     override fun execute(params: Param): Completable = balanceDataSource.saveBalance(params.balance)
 }
 
-class Param(val balance: Map<Currency, Double>)
+class Param(val balance: Map<Currency, BigDecimal>)
